+++ src/isa/toy-isa/inst.h
@@ -159,7 +159,7 @@ class toy_isa_inst: public inst {
   int get_num_operands() const {return _isa.num_operands[_opcode];}
   int get_insn_num_regs() const {return _isa.insn_num_regs[_opcode];}
   int get_opcode_type() const {return _isa.opcode_type[_opcode];}
-  int interpret(const vector<inst*> &instptr_list, prog_state &ps, int input) const;
+  int64_t interpret(const vector<inst*> &instptr_list, prog_state &ps, int input) const;
   // smt
   z3::expr smt_inst(smt_var& sv) const;
   z3::expr smt_inst_jmp(smt_var& sv) const;
